renamed sessions store to realtimesessionsstore
